pyproject.toml
[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[project]
name = "uup-builder"
version = "0.1.0"
description = "Python Module for building an iso from uup dump files."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "requests>=2.32.5",
    "uup-dump-api-py>=0.1.5",
]

[project.optional-dependencies]
pretty = ["rich>=13.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
uup-builder = "uup_builder.cli:main"

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
